pyproject.toml
[project]
name = "ontology-intel-hub"
version = "0.1.0"
description = "Intel hub wrapper around TrendRadar outputs"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "jinja2>=3.1,<4.0",
  "numpy>=1.26,<3",
  "pydantic>=2.11,<3.0",
  "pyyaml>=6.0,<7.0",
  "scikit-learn>=1.3,<2",
  "uvicorn>=0.35,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps"]
